+++ src/Context/BizAwareInterface.php
@@ -1,0 +1,9 @@
+<?php
+namespace Codeages\PhalconBiz;
+
+use Codeages\Biz\Framework\Context\Biz;
+
+interface BizAwareInterface
+{
+    public function setBiz(Biz $biz);
+}