+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased]
 
 * 删除`Codeages\Biz\Framework\Context\CurrentUserInterface`接口，新增 `Codeages\Biz\Framework\Context\CurrentUser`类。
-
+* 新增`Codeages\Biz\Framework\Context\BizAwareInterface`接口。
 
 
 ## [0.8.4] - 2017-09-14